rpc: sunset the old experimental rest service
