+++ crates/sui-rpc-api/src/service/health.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+use axum::extract::{Query, State};
 use std::time::Duration;
 use std::time::SystemTime;
 
@@ -33,5 +34,25 @@ impl RpcService {
         }
 
         Ok(())
+    }
+}
+
+#[derive(Debug, serde::Serialize, serde::Deserialize)]
+pub struct Threshold {
+    /// The threshold, or delta, between the server's system time and the timestamp in the most
+    /// recently executed checkpoint for which the server is considered to be healthy.
+    ///
+    /// If not provided, the server will be considered healthy if it can simply fetch the latest
+    /// checkpoint from its store.
+    pub threshold_seconds: Option<u32>,
+}
+
+pub async fn health(
+    Query(Threshold { threshold_seconds }): Query<Threshold>,
+    State(state): State<RpcService>,
+) -> impl axum::response::IntoResponse {
+    match state.health_check(threshold_seconds) {
+        Ok(()) => (axum::http::StatusCode::OK, "up"),
+        Err(_) => (axum::http::StatusCode::SERVICE_UNAVAILABLE, "down"),
     }
 }
